# improvements to `(i)starts_with` and `(i)ends_with` performance