@alamb 2024-07-27
Here is my next run


```
++ critcmp master starts_with-ends_with-improvements
group                                              master                                 starts_with-ends_with-improvements
-----                                              ------                                 ----------------------------------
ilike_utf8 scalar complex                          1.00    301.7±1.10µs        ? ?/sec    1.00    301.5±1.04µs        ? ?/sec
ilike_utf8 scalar contains                         1.00   1554.5±6.20µs        ? ?/sec    1.03   1595.8±4.49µs        ? ?/sec
ilike_utf8 scalar ends with                        1.00    218.5±0.39µs        ? ?/sec    1.13    247.1±4.53µs        ? ?/sec
ilike_utf8 scalar equals                           1.16    248.8±0.40µs        ? ?/sec    1.00    214.8±0.66µs        ? ?/sec
ilike_utf8 scalar starts with                      1.00    279.4±0.36µs        ? ?/sec    1.07    298.5±0.46µs        ? ?/sec
ilike_utf8_scalar_dyn dictionary[10] string[4])    1.00     88.1±0.14µs        ? ?/sec    1.00     88.2±0.18µs        ? ?/sec
like_utf8 scalar complex                           1.01    283.4±0.82µs        ? ?/sec    1.00    282.0±0.96µs        ? ?/sec
like_utf8 scalar contains                          1.00    347.6±0.52µs        ? ?/sec    1.00    347.4±0.76µs        ? ?/sec
like_utf8 scalar ends with                         1.41    219.4±0.55µs        ? ?/sec    1.00    155.3±2.99µs        ? ?/sec
like_utf8 scalar equals                            1.00    217.6±0.53µs        ? ?/sec    1.00    217.9±0.73µs        ? ?/sec
like_utf8 scalar starts with                       1.34    232.8±0.29µs        ? ?/sec    1.00    173.5±0.30µs        ? ?/sec
like_utf8_scalar_dyn dictionary[10] string[4])     1.00     88.1±0.17µs        ? ?/sec    1.00     88.1±0.14µs        ? ?/sec
like_utf8view scalar complex                       1.00    531.3±2.04ms        ? ?/sec    1.00    531.0±2.43ms        ? ?/sec
like_utf8view scalar contains                      1.00    378.6±0.36ms        ? ?/sec    1.01    380.6±1.38ms        ? ?/sec
like_utf8view scalar ends with                     1.13     59.6±0.24ms        ? ?/sec    1.00     52.7±0.23ms        ? ?/sec
like_utf8view scalar equals                        1.00     37.0±0.47ms        ? ?/sec    1.00     36.9±0.09ms        ? ?/sec
like_utf8view scalar starts with                   1.06     60.0±0.48ms        ? ?/sec    1.00     56.8±0.23ms        ? ?/sec
nilike_utf8 scalar complex                         1.00    301.9±1.19µs        ? ?/sec    1.00    300.4±2.48µs        ? ?/sec
nilike_utf8 scalar contains                        1.00   1551.3±4.82µs        ? ?/sec    1.03   1594.3±4.71µs        ? ?/sec
nilike_utf8 scalar ends with                       1.00    219.7±3.41µs        ? ?/sec    1.12    246.3±0.86µs        ? ?/sec
nilike_utf8 scalar equals                          1.16    248.8±0.53µs        ? ?/sec    1.00    214.5±0.33µs        ? ?/sec
nilike_utf8 scalar starts with                     1.00    279.6±1.02µs        ? ?/sec    1.07    298.4±0.46µs        ? ?/sec
nlike_utf8 scalar complex                          1.01    283.2±1.05µs        ? ?/sec    1.00    281.1±1.19µs        ? ?/sec
nlike_utf8 scalar contains                         1.00    347.5±0.63µs        ? ?/sec    1.00    347.2±0.49µs        ? ?/sec
nlike_utf8 scalar ends with                        1.42    219.7±1.25µs        ? ?/sec    1.00    154.8±0.24µs        ? ?/sec
nlike_utf8 scalar equals                           1.00    217.7±0.45µs        ? ?/sec    1.00    217.7±0.47µs        ? ?/sec
nlike_utf8 scalar starts with                      1.34    233.0±0.85µs        ? ?/sec    1.00    173.6±0.48µs        ? ?/sec
```

BTW I am running this on a `c2-standard-8` GCP instance

Script
```shell
pushd ~/arrow-rs

#git remote add samuelcolvin https://github.com/samuelcolvin/arrow-rs.git
git fetch -p samuelcolvin
BENCH_COMMAND="cargo bench -p arrow --bench comparison_kernels -F test_utils"
BENCH_FILTER="like"
REPO_NAME="samuelcolvin"
BRANCH_NAME="starts_with-ends_with-improvements"

# remove old test runs
rm -rf target/criterion/

git checkout $BRANCH_NAME
git reset --hard "$REPO_NAME/$BRANCH_NAME"

# Run on test branch
$BENCH_COMMAND -- --save-baseline ${BRANCH_NAME} ${BENCH_FILTER}

# Run on master
MERGE_BASE=$(git merge-base HEAD apache/master)
echo "** Comparing to ${MERGE_BASE}"

git checkout ${MERGE_BASE}
$BENCH_COMMAND -- --save-baseline master  ${BENCH_FILTER}

critcmp master ${BRANCH_NAME}

popd
```